refactor: move ruffus related code to one file
