+++ src/ocrmypdf/_validation.py
@@ -16,33 +16,20 @@
 # You should have received a copy of the GNU General Public License
 # along with OCRmyPDF.  If not, see <http://www.gnu.org/licenses/>.
 
-import atexit
+
 import logging
 import os
-import re
+
 import sys
 import textwrap
 from pathlib import Path
-from tempfile import mkdtemp
 
 import PIL
-import ruffus.cmdline as cmdline
-import ruffus.proxy_logger as proxy_logger
-import ruffus.ruffus_exceptions as ruffus_exceptions
 
 from . import VERSION
-from . import exceptions as ocrmypdf_exceptions
-from ._jobcontext import JobContext, JobContextManager, cleanup_working_files
-from ._pipeline import build_pipeline
+
 from ._unicodefun import verify_python3_env
-from .exceptions import (
-    BadArgsError,
-    ExitCode,
-    ExitCodeException,
-    InputFileError,
-    MissingDependencyError,
-    OutputFileAccessError,
-)
+
 from .exec import (
     ghostscript,
     jbig2enc,
@@ -52,8 +39,14 @@ from .exec import (
     unpaper,
     pngquant,
 )
-from .helpers import available_cpu_count, is_file_writable, re_symlink
-from .pdfa import file_claims_pdfa
+from .helpers import is_file_writable, re_symlink
+from .exceptions import (
+    BadArgsError,
+    ExitCode,
+    InputFileError,
+    MissingDependencyError,
+    OutputFileAccessError,
+)
 
 # -------------
 # External dependencies
@@ -64,9 +57,9 @@ HOCR_OK_LANGS = frozenset(['eng', 'deu', 'spa', 'ita', 'por'])
 def complain(message):
     print(*textwrap.wrap(message), file=sys.stderr)
 
+
 # --------
 # Critical environment tests
-
 verify_python3_env()
 
 
@@ -305,102 +298,6 @@ def logging_factory(logger_name, logger_args):
     return root_logger
 
 
-def cleanup_ruffus_error_message(msg):
-    msg = re.sub(r'\s+', r' ', msg)
-    msg = re.sub(r"\((.+?)\)", r'\1', msg)
-    msg = msg.strip()
-    return msg
-
-
-def do_ruffus_exception(ruffus_five_tuple, options, log):
-    """Replace the elaborate ruffus stack trace with a user friendly
-    description of the error message that occurred."""
-    exit_code = None
-
-    _task_name, _job_name, exc_name, exc_value, exc_stack = ruffus_five_tuple
-
-    if isinstance(exc_name, type):
-        # ruffus is full of mystery... sometimes (probably when the process
-        # group leader is killed) exc_name is the class object of the exception,
-        # rather than a str. So reach into the object and get its name.
-        exc_name = exc_name.__name__
-
-    if exc_name.startswith('ocrmypdf.exceptions.'):
-        base_exc_name = exc_name.replace('ocrmypdf.exceptions.', '')
-        exc_class = getattr(ocrmypdf_exceptions, base_exc_name)
-        exit_code = getattr(exc_class, 'exit_code', ExitCode.other_error)
-        try:
-            if isinstance(exc_value, exc_class):
-                exc_msg = str(exc_value)
-            elif isinstance(exc_value, str):
-                exc_msg = exc_value
-            else:
-                exc_msg = str(exc_class())
-        except Exception:
-            exc_msg = "Unknown"
-
-    if exc_name in ('builtins.SystemExit', 'SystemExit'):
-        match = re.search(r"\.(.+?)\)", exc_value)
-        exit_code_name = match.groups()[0]
-        exit_code = getattr(ExitCode, exit_code_name, 'other_error')
-    elif exc_name == 'ruffus.ruffus_exceptions.MissingInputFileError':
-        log.error(cleanup_ruffus_error_message(exc_value))
-        exit_code = ExitCode.input_file
-    elif exc_name in ('builtins.KeyboardInterrupt', 'KeyboardInterrupt'):
-        # We have to print in this case because the log daemon might be toast
-        print("Interrupted by user", file=sys.stderr)
-        exit_code = ExitCode.ctrl_c
-    elif exc_name == 'subprocess.CalledProcessError':
-        # It's up to the subprocess handler to report something useful
-        msg = "Error occurred while running this command:"
-        log.error(msg + '\n' + exc_value)
-        exit_code = ExitCode.child_process_error
-    elif exc_name.startswith('ocrmypdf.exceptions.'):
-        if exc_msg:
-            log.error(exc_msg)
-    elif exc_name == 'PIL.Image.DecompressionBombError':
-        msg = cleanup_ruffus_error_message(exc_value)
-        msg += (
-            "\nUse the --max-image-mpixels argument to set increase the "
-            "maximum number of megapixels to accept."
-        )
-        log.error(msg)
-        exit_code = ExitCode.input_file
-
-    if exit_code is not None:
-        return exit_code
-
-    if not options.verbose:
-        log.error(exc_stack)
-    return ExitCode.other_error
-
-
-def traverse_ruffus_exception(exceptions, options, log):
-    """Traverse a RethrownJobError and output the exceptions
-
-    Ruffus presents exceptions as 5 element tuples. The RethrownJobException
-    has a list of exceptions like
-        e.job_exceptions = [(5-tuple), (5-tuple), ...]
-
-    ruffus < 2.7.0 had a bug with exception marshalling that would give
-    different output whether the main or child process raised the exception.
-    We no longer support this.
-
-    Attempting to log the exception itself will re-marshall it to the logger
-    which is normally running in another process. It's better to avoid re-
-    marshalling.
-
-    The exit code will be based on this, even if multiple exceptions occurred
-    at the same time."""
-
-    exit_codes = []
-    for exc in exceptions:
-        exit_code = do_ruffus_exception(exc, options, log)
-        exit_codes.append(exit_code)
-
-    return exit_codes[0]  # Multiple codes are rare so take the first one
-
-
 def check_closed_streams(options):
     """Work around Python issue with multiprocessing forking on closed streams
 
@@ -516,10 +413,7 @@ def check_requested_output_file(options, _log):
             raise BadArgsError()
     elif not is_file_writable(options.output_file):
         _log.error(
-            "Output file location ("
-            + options.output_file
-            + ") "
-            + "is not a writable file."
+            "Output file location (" + options.output_file + ") is not a writable file."
         )
         raise OutputFileAccessError()
 
@@ -594,109 +488,3 @@ def check_dependency_versions(options, log):
         version_checker=qpdf.version,
         need_version='8.0.2',
     )
-
-
-def run_pipeline(options):
-    options.verbose_abbreviated_path = 1
-    if os.environ.get('_OCRMYPDF_THREADS'):
-        options.use_threads = True
-
-    if not check_closed_streams(options):
-        return ExitCode.bad_args
-
-    logger_args = {'verbose': options.verbose, 'quiet': options.quiet}
-
-    _log, _log_mutex = proxy_logger.make_shared_logger_and_proxy(
-        logging_factory, __name__, logger_args
-    )
-    preamble(_log)
-    check_code = check_options(options, _log)
-    if check_code != ExitCode.ok:
-        return check_code
-    check_dependency_versions(options, _log)
-
-    # Any changes to options will not take effect for options that are already
-    # bound to function parameters in the pipeline. (For example
-    # options.input_file, options.pdf_renderer are already bound.)
-    if not options.jobs:
-        options.jobs = available_cpu_count()
-
-    # Performance is improved by setting Tesseract to single threaded. In tests
-    # this gives better throughput than letting a smaller number of Tesseract
-    # jobs run multithreaded. Same story for pngquant. Tess <4 ignores this
-    # variable, but harmless to set if ignored.
-    os.environ.setdefault('OMP_THREAD_LIMIT', '1')
-
-    check_environ(options, _log)
-    if os.environ.get('PYTEST_CURRENT_TEST'):
-        os.environ['_OCRMYPDF_TEST_INFILE'] = options.input_file
-
-    try:
-        work_folder = mkdtemp(prefix="com.github.ocrmypdf.")
-        options.history_file = os.path.join(work_folder, 'ruffus_history.sqlite')
-        start_input_file = os.path.join(work_folder, 'origin')
-
-        check_input_file(options, _log, start_input_file)
-        check_requested_output_file(options, _log)
-
-        manager = JobContextManager()
-        manager.register('JobContext', JobContext)  # pylint: disable=no-member
-        manager.start()
-
-        context = manager.JobContext()  # pylint: disable=no-member
-        context.set_options(options)
-        context.set_work_folder(work_folder)
-
-        build_pipeline(options, work_folder, _log, context)
-        atexit.register(cleanup_working_files, work_folder, options)
-        if hasattr(os, 'nice'):
-            os.nice(5)
-        cmdline.run(options)
-    except ruffus_exceptions.RethrownJobError as e:
-        if options.verbose:
-            _log.debug(str(e))  # stringify exception so logger doesn't have to
-        exceptions = e.job_exceptions
-        exitcode = traverse_ruffus_exception(exceptions, options, _log)
-        if exitcode is None:
-            _log.error("Unexpected ruffus exception: " + str(e))
-            _log.error(repr(e))
-            return ExitCode.other_error
-        return exitcode
-    except ExitCodeException as e:
-        return e.exit_code
-    except Exception as e:
-        _log.error(str(e))
-        return ExitCode.other_error
-
-    if options.flowchart:
-        _log.info(f"Flowchart saved to {options.flowchart}")
-        return ExitCode.ok
-    elif options.output_file == '-':
-        _log.info("Output sent to stdout")
-    elif os.path.samefile(options.output_file, os.devnull):
-        pass  # Say nothing when sending to dev null
-    else:
-        if options.output_type.startswith('pdfa'):
-            pdfa_info = file_claims_pdfa(options.output_file)
-            if pdfa_info['pass']:
-                msg = f"Output file is a {pdfa_info['conformance']} (as expected)"
-                _log.info(msg)
-            else:
-                msg = f"Output file is okay but is not PDF/A (seems to be {pdfa_info['conformance']})"
-                _log.warning(msg)
-                return ExitCode.pdfa_conversion_failed
-        if not qpdf.check(options.output_file, _log):
-            _log.warning('Output file: The generated PDF is INVALID')
-            return ExitCode.invalid_output_pdf
-
-        report_output_file_size(options, _log, start_input_file, options.output_file)
-
-    pdfinfo = context.get_pdfinfo()
-    if options.verbose:
-        from pprint import pformat
-
-        _log.debug(pformat(pdfinfo))
-
-    log_page_orientations(pdfinfo, _log)
-
-    return ExitCode.ok
