+++ tests/test_unpaper.py
@@ -15,15 +15,13 @@
 # You should have received a copy of the GNU General Public License
 # along with OCRmyPDF.  If not, see <http://www.gnu.org/licenses/>.
 
-import argparse
 from os import fspath
-from pathlib import Path
 from unittest.mock import MagicMock, patch
 
 import pytest
 
 from ocrmypdf.__main__ import parser
-from ocrmypdf.run import check_options
+from ocrmypdf._validation import check_options
 from ocrmypdf.exceptions import ExitCode
 from ocrmypdf.exec import unpaper
